Creation of modules for the framework and App
Creation and refactor of classes for the framework and Application activity
