--- src/app/Main.java
@@ -1,11 +1,0 @@
-package app;
-
-import framework.Framework;
-import framework.annotation.App;
-
-@App("app")
-public class Main {
-    public static void main(String[] args){
-        Framework.run(Main.class);
-    }
-}
