--- src/framework/annotation/Autowire.java
@@ -1,8 +1,0 @@
-package framework.annotation;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Autowire {
-}
