--- src/framework/annotation/Body.java
@@ -1,4 +1,0 @@
-package framework.annotation;
-
-public @interface Body {
-}
