--- src/framework/annotation/Component.java
@@ -1,4 +1,0 @@
-package framework.annotation;
-
-public @interface Component {
-}
