--- src/framework/annotation/DELETE.java
@@ -1,5 +1,0 @@
-package framework.annotation;
-
-public @interface DELETE {
-    String value();
-}
