--- src/framework/annotation/PUT.java
@@ -1,5 +1,0 @@
-package framework.annotation;
-
-public @interface PUT {
-    String value();
-}
