--- src/framework/annotation/PathParam.java
@@ -1,5 +1,0 @@
-package framework.annotation;
-
-public @interface PathParam {
-    String value();
-}
